refactor: Delete last element from linked list
The code changes in the `fim lista encadeada.c` file refactor the linked list by adding a new function `delete_final` that deletes the last element from the list. This function iterates through the list to find the second-to-last element, updates its `next` pointer to `NULL`, and frees the memory of the last element.

Note: The commit message follows the established convention of using a prefix (refactor) to indicate code refactoring.