feat: Add file I/O functionality and line count feature
The code changes in the `arquivo3.c` file add file I/O functionality and a line count feature. The user is prompted to enter the name of a file, which is then opened and read line by line to count the number of lines. The line count is then displayed to the user. This feature enhances the functionality of the program and provides useful information to the user.

Note: The commit message follows the established convention of using a prefix (feat) to indicate a new feature.